requirements.txt
langchain-community 
langchain-text-splitters 
langchain-groq
langchain-huggingface 
langchain-chroma 
sentence-transformers
gradio 
python-dotenv 
pypdf